test(vetting): one declaration shape and one value type for every reference table
Three tables were still raw std::unordered_map (the two firstorder pyradiomics
tables and intensity_histogram_2d_ibsi_ref_vals) and five were ref_vals_map<float>
(the 3D glcm/gldm/glrlm/glszm/ngtdm regression tables). agrees_gt and ASSERT_NEAR
both compare in double, so a float table narrowed its literals only to widen them
again at the assertion; the values move by at most ~6e-8 relative against 10% bands.

check_test_names.py now rejects a reference table declared as a raw container, so
the set stays identifiable by type rather than by the words in a name. The self-test
plants one whose name and location both conform, leaving the type as the only
defect. 61 tables inspected, 0 raw.

SPEC 6.3.1 records both rules and why.

+++ tests/test_3d_glszm_regression.h
@@ -9,7 +9,7 @@
 
 // dig. phantom values for intensity based features
 // Calculated at 100 grey levels
-static ref_vals_map<float> glszm_3d_regression_ref_vals{
+static ref_vals_map<double> glszm_3d_regression_ref_vals{
     {"3GLSZM_SAE",  0.6},
     {"3GLSZM_LAE",  1377.1},
     {"3GLSZM_LGLZE",    0.0005},
